backend/requirements.txt
fastapi==0.136.1
itsdangerous==2.2.0
python-dotenv==1.2.2
slowapi==0.1.9
uvicorn==0.46.0
sqlalchemy[asyncio]==2.0.49
asyncpg==0.31.0
bcrypt==5.0.0
PyJWT==2.12.1
pydantic[email]==2.13.4
